requirements.txt
netmiko
pycryptodome
cryptography
colorama
psutil
